Move Qubit to new file
